Make a common class for DFT runners
