+++ component-tests/pwdft/run_test.py
@@ -15,8 +15,6 @@ from parsl.providers import PBSProProvider
 from parsl.launchers import SimpleLauncher
 
 from mofa.model import MOFRecord
-from mofa.scoring.geometry import LatticeParameterChange
-from mofa.simulation.cp2k import compute_partial_charges
 from mofa.utils.conversions import write_to_string
 
 
@@ -31,7 +29,7 @@ def test_function(strc: MOFRecord, invocation: str, steps: int) -> tuple[float, 
         - Runtime (s)
         - MD trajectory
     """
-    from mofa.simulation.pwdft import PWDFTRunner
+    from mofa.simulation.dft.pwdft import PWDFTRunner
     from time import perf_counter
     from pathlib import Path
     import shutil
