+++ docs/source/simulation.rst
@@ -1,0 +1,15 @@
+Simulation (``mofa.simulation``)
+================================
+
+.. automodule:: mofa.simulation
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+DFT (``m.simulation.dft``)
+--------------------------
+
+.. automodule:: mofa.simulation.dft
+   :members:
+   :undoc-members:
+   :show-inheritance:
