+++ mofa/simulation/dft/files/cp2k-default-template.inp
