+++ scripts/0_initialize-database/1_compute-partial-charges.py
@@ -16,7 +16,7 @@ from parsl.providers import PBSProProvider
 from parsl.launchers import SimpleLauncher
 
 from mofa.model import MOFRecord
-from mofa.simulation.cp2k import compute_partial_charges
+from mofa.simulation.dft import compute_partial_charges
 from mofa.simulation.mace import MACERunner
 from mofa.utils.conversions import write_to_string
 
@@ -32,7 +32,7 @@ def run_cp2k(strc: MOFRecord, cp2k_invocation: str, steps: int) -> tuple[float, 
         - Runtime (s)
         - MD trajectory
     """
-    from mofa.simulation.cp2k import CP2KRunner
+    from mofa.simulation.dft.cp2k import CP2KRunner
     from time import perf_counter
     from pathlib import Path
