+++ tests/simulation/test_cp2k.py
@@ -4,7 +4,8 @@ import shutil
 from pytest import mark
 
 from mofa.model import MOFRecord
-from mofa.simulation.cp2k import CP2KRunner, compute_partial_charges
+from mofa.simulation.dft.cp2k import CP2KRunner
+from mofa.simulation.dft import compute_partial_charges
 from mofa.utils.conversions import write_to_string
 
 IN_GITHUB_ACTIONS = os.getenv("GITHUB_ACTIONS") == "true"
@@ -14,14 +15,17 @@ IN_GITHUB_ACTIONS = os.getenv("GITHUB_ACTIONS") == "true"
 def test_cp2k_single(cif_name, cif_dir, tmpdir):
     # Make a CP2k simulator that reads and writes to a temporary directory
     runner = CP2KRunner(
-        cp2k_invocation="cp2k_shell",
+        dft_cmd="cp2k_shell",
         close_cp2k=False,
     )
 
     test_file = cif_dir / f'{cif_name}.cif'
     record = MOFRecord.from_file(test_file)
     record.md_trajectory['uff'] = [write_to_string(record.atoms, 'vasp')]
-    _, cp2k_path = runner.run_single_point(record, structure_source=('uff', -1))
+    atoms, cp2k_path = runner.run_single_point(record, structure_source=('uff', -1))
+
+    assert atoms.calc is not None
+    assert atoms.get_forces().shape == (len(atoms), 3)  # Assure it at least gives some numbers
 
     charged_mof = compute_partial_charges(cp2k_path, threads=2)
     assert charged_mof.arrays["q"].shape[0] == charged_mof.arrays["positions"].shape[0]
@@ -41,14 +45,15 @@ def test_cp2k_optimize(cif_name, cif_dir, tmpdir):
 
     # Make a CP2k simulator that reads and writes to a temporary directory
     runner = CP2KRunner(
-        cp2k_invocation="cp2k_shell",
+        dft_cmd="cp2k_shell",
         close_cp2k=True
     )
 
     test_file = cif_dir / f'{cif_name}.cif'
     record = MOFRecord.from_file(test_file)
     record.md_trajectory['uff'] = [write_to_string(record.atoms, 'vasp')]
-    _, cp2k_path = runner.run_optimization(record, steps=2, fmax=0.1)
+    atoms, cp2k_path = runner.run_optimization(record, steps=2, fmax=0.1)
+    assert atoms.get_potential_energy() is not None
     assert cp2k_path.exists()
     assert cp2k_path.is_absolute()
     assert 'opt' in cp2k_path.name
