+++ tests/simulation/test_pwdft.py
@@ -1,10 +1,10 @@
 import os
 import shutil
-import subprocess
+from shutil import which
 
 from pytest import mark
 from mofa.model import MOFRecord
-from mofa.simulation.pwdft import PWDFTRunner
+from mofa.simulation.dft.pwdft import PWDFTRunner
 from mofa.utils.conversions import write_to_string
 
 IN_GITHUB_ACTIONS = os.getenv("GITHUB_ACTIONS") == "true"
@@ -12,18 +12,15 @@ IN_GITHUB_ACTIONS = os.getenv("GITHUB_ACTIONS") == "true"
 
 def has_pwdft():
     """Check if pwdft binary is available in PATH"""
-    try:
-        subprocess.run(["pwdft", "--version"], capture_output=True, check=True)
-        return True
-    except (subprocess.SubprocessError, FileNotFoundError):
-        return False
+    return which('pwdft') is not None
 
 
+@mark.slow
 @mark.skipif(not has_pwdft(), reason="PWDFT binary not found in PATH")
 @mark.parametrize("cif_name", ["hMOF-0"])
 def test_pwdft_single(cif_name, cif_dir, tmpdir):
     # Make a PWDFT simulator that reads and writes to a temporary directory
-    runner = PWDFTRunner()
+    runner = PWDFTRunner(dft_cmd="mpiexec -n 12 pwdft")
 
     test_file = cif_dir / f"{cif_name}.cif"
     record = MOFRecord.from_file(test_file)
@@ -38,13 +35,14 @@ def test_pwdft_single(cif_name, cif_dir, tmpdir):
     assert atoms.get_potential_energy() is not None
 
 
+@mark.slow
 @mark.skipif(IN_GITHUB_ACTIONS, reason="Too expensive for CI")
 @mark.skipif(not has_pwdft(), reason="PWDFT binary not found in PATH")
 @mark.parametrize("cif_name", ["hMOF-0"])
 def test_pwdft_optimize(cif_name, cif_dir, tmpdir):
     shutil.rmtree("pwdft-runs", ignore_errors=True)
 
-    runner = PWDFTRunner()
+    runner = PWDFTRunner(dft_cmd="mpiexec -n 12 pwdft")
 
     test_file = cif_dir / f"{cif_name}.cif"
     record = MOFRecord.from_file(test_file)
@@ -66,7 +64,7 @@ def test_pwdft_optimize(cif_name, cif_dir, tmpdir):
 @mark.parametrize("level", ["default"])
 def test_pwdft_options(level, cif_dir):
     """Test that different PWDFT options work"""
-    runner = PWDFTRunner()
+    runner = PWDFTRunner(dft_cmd="mpirun -n 12 pwdft")
 
     test_file = cif_dir / "hMOF-0.cif"
     record = MOFRecord.from_file(test_file)
